Move ParameterizedTypeName
