+++ src/main/java/com/squareup/kotlinpoet/ParameterizedTypeName.kt
@@ -17,148 +17,24 @@
 
 package com.squareup.kotlinpoet
 
-import java.lang.reflect.Modifier
 import java.lang.reflect.ParameterizedType
-import java.lang.reflect.Type
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.reflect.KTypeParameter
-import kotlin.reflect.KTypeProjection
-import kotlin.reflect.KVariance
 
-class ParameterizedTypeName internal constructor(
-  private val enclosingType: TypeName?,
-  val rawType: ClassName,
-  typeArguments: List<TypeName>,
-  nullable: Boolean = false,
-  annotations: List<AnnotationSpec> = emptyList()
-) : TypeName(nullable, annotations) {
-  val typeArguments = typeArguments.toImmutableList()
+fun ParameterizedTypeName.plusParameter(typeArgument: TypeName) = ParameterizedTypeName(enclosingType, rawType,
+    typeArguments + typeArgument, nullable, annotations)
 
-  init {
-    require(typeArguments.isNotEmpty() || enclosingType != null) {
-      "no type arguments: $rawType"
-    }
-  }
+fun ParameterizedTypeName.plusParameter(typeArgument: KClass<*>) = plusParameter(typeArgument.asClassName())
 
-  override fun asNullable()
-      = ParameterizedTypeName(enclosingType, rawType, typeArguments, true, annotations)
+fun ParameterizedTypeName.plusParameter(typeArgument: Class<*>) = plusParameter(typeArgument.asClassName())
 
-  override fun asNonNull()
-      = ParameterizedTypeName(enclosingType, rawType, typeArguments, false, annotations)
-
-  override fun annotated(annotations: List<AnnotationSpec>) = ParameterizedTypeName(
-      enclosingType, rawType, typeArguments, nullable, this.annotations + annotations)
-
-  override fun withoutAnnotations()
-      = ParameterizedTypeName(enclosingType, rawType, typeArguments, nullable)
-
-  fun plusParameter(typeArgument: TypeName) = ParameterizedTypeName(enclosingType, rawType,
-      typeArguments + typeArgument, nullable, annotations)
-
-  fun plusParameter(typeArgument: KClass<*>) = plusParameter(typeArgument.asClassName())
-
-  fun plusParameter(typeArgument: Class<*>) = plusParameter(typeArgument.asClassName())
-
-  override fun emit(out: CodeWriter): CodeWriter {
-    if (enclosingType != null) {
-      enclosingType.emitAnnotations(out)
-      enclosingType.emit(out)
-      out.emit("." + rawType.simpleName)
-    } else {
-      rawType.emitAnnotations(out)
-      rawType.emit(out)
-    }
-    if (typeArguments.isNotEmpty()) {
-      out.emit("<")
-      typeArguments.forEachIndexed { index, parameter ->
-        if (index > 0) out.emit(", ")
-        parameter.emitAnnotations(out)
-        parameter.emit(out)
-        parameter.emitNullable(out)
-      }
-      out.emit(">")
-    }
-    return out
-  }
-
-  /**
-   * Returns a new [ParameterizedTypeName] instance for the specified `name` as nested inside this
-   * class, with the specified `typeArguments`.
-   */
-  fun nestedClass(name: String, typeArguments: List<TypeName>)
-      = ParameterizedTypeName(this, rawType.nestedClass(name), typeArguments)
-
-  companion object {
-    /** Returns a parameterized type, applying `typeArguments` to `this`.  */
-    @JvmStatic @JvmName("get") fun ClassName.parameterizedBy(vararg typeArguments: TypeName)
-        = ParameterizedTypeName(null, this, typeArguments.toList())
-
-    /** Returns a parameterized type, applying `typeArguments` to `this`.  */
-    @JvmStatic @JvmName("get") fun KClass<*>.parameterizedBy(vararg typeArguments: KClass<*>)
-        = ParameterizedTypeName(null, asClassName(), typeArguments.map { it.asTypeName() })
-
-    /** Returns a parameterized type, applying `typeArguments` to `this`.  */
-    @JvmStatic @JvmName("get") fun Class<*>.parameterizedBy(vararg typeArguments: Type) =
-        ParameterizedTypeName(null, asClassName(), typeArguments.map { it.asTypeName() })
-
-    /** Returns a parameterized type, applying `typeArgument` to `this`.  */
-    @JvmStatic @JvmName("get") fun ClassName.plusParameter(typeArgument: TypeName) =
-        parameterizedBy(typeArgument)
-
-    /** Returns a parameterized type, applying `typeArgument` to `this`.  */
-    @JvmStatic @JvmName("get") fun KClass<*>.plusParameter(typeArgument: KClass<*>) =
-        parameterizedBy(typeArgument)
-
-    /** Returns a parameterized type, applying `typeArgument` to `this`.  */
-    @JvmStatic @JvmName("get") fun Class<*>.plusParameter(typeArgument: Class<*>) =
-        parameterizedBy(typeArgument)
-
-    /** Returns a parameterized type equivalent to `type`.  */
-    internal fun get(
-      type: ParameterizedType,
-      map: MutableMap<Type, TypeVariableName>
-    ): ParameterizedTypeName {
-      val rawType = (type.rawType as Class<*>).asClassName()
-      val ownerType = if (type.ownerType is ParameterizedType
-          && !Modifier.isStatic((type.rawType as Class<*>).modifiers))
-        type.ownerType as ParameterizedType else
-        null
-
-      val typeArguments = type.actualTypeArguments.map { TypeName.get(it, map = map) }
-      return if (ownerType != null)
-        get(ownerType, map = map).nestedClass(rawType.simpleName, typeArguments) else
-        ParameterizedTypeName(null, rawType, typeArguments)
-    }
-
-    /** Returns a type name equivalent to type with given list of type arguments.  */
-    internal fun get(type: KClass<*>, nullable: Boolean, typeArguments: List<KTypeProjection>): TypeName {
-      if (typeArguments.isEmpty()) {
-        return type.asTypeName()
-      }
-
-      val effectiveType = if (type.java.isArray) Array<Unit>::class else type
-      val enclosingClass = type.java.enclosingClass?.kotlin
-
-      return ParameterizedTypeName(
-          enclosingClass?.let {
-            get(it, false, typeArguments.drop(effectiveType.typeParameters.size))
-          },
-          effectiveType.asTypeName(),
-          typeArguments.take(effectiveType.typeParameters.size).map { (paramVariance, paramType) ->
-            val typeName = paramType?.asTypeName() ?: return@map WildcardTypeName.STAR
-            when (paramVariance) {
-              null -> WildcardTypeName.STAR
-              KVariance.INVARIANT -> typeName
-              KVariance.IN -> WildcardTypeName.supertypeOf(typeName)
-              KVariance.OUT -> WildcardTypeName.subtypeOf(typeName)
-            }
-          },
-          nullable,
-          effectiveType.annotations.map { AnnotationSpec.get(it) })
-    }
-  }
-}
+/**
+ * Returns a new [ParameterizedTypeName] instance for the specified `name` as nested inside this
+ * class, with the specified `typeArguments`.
+ */
+fun ParameterizedTypeName.nestedClass(name: String, typeArguments: List<TypeName>)
+    = ParameterizedTypeName(this, rawType.nestedClass(name), typeArguments)
 
 /** Returns a parameterized type equivalent to `type`.  */
 @JvmName("get")
